+++ src/torchrunx/__init__.py
@@ -1,4 +1,5 @@
 from .launcher import Launcher, launch
+from .log_utils import DefaultLogSpec, LogSpec
 from .slurm import slurm_hosts, slurm_workers
 
-__all__ = ["Launcher", "launch", "slurm_hosts", "slurm_workers"]
+__all__ = ["Launcher", "launch", "slurm_hosts", "slurm_workers", "LogSpec", "DefaultLogSpec"]
